requirements.txt
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-bootstrap-components==1.0.1
pandas==1.3.4
plotly==5.3.1
requests==2.26.0
PyYAML
